+++ Bot/Bot/Indicators/Interfaces/IMovingAverageCrossover.cs
@@ -5,7 +5,7 @@ namespace Bot.Indicators.Interfaces
 {
 	public interface IMovingAverageCrossover : ISimpleValueIndicator<PositionType>
 	{
-		public double ShortMa { get; }
-		public double LongMa { get; }
+		public decimal ShortMa { get; }
+		public decimal LongMa { get; }
 	}
 }
